ResumeCoverLetterMaker/pyproject.toml
[project]
name = "jobapplicationautomation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "google-genai>=1.0.0",
    "jinja2>=3.1.0",
    "lxml>=5.0.0",
    "playwright>=1.40.0",
    "pypdf>=4.0.0",
    "pyyaml>=6.0.1",
    "reportlab>=4.0.0",
    "requests>=2.31.0",
    "xhtml2pdf>=0.2.11",
]
